bugz

+++ pizza_cutter_sims/condor_exec.py
@@ -102,13 +102,16 @@ def _nanny_function(
             return
 
         subids = list(exec._nanny_subids[nanny_id])
+        if DEBUG:
+            print("getting job statuses", flush=True)
         statuses = _get_all_job_statuses([
             exec._nanny_subids[nanny_id][subid][0]
             for subid in subids
             if exec._nanny_subids[nanny_id][subid][0] is not None
         ])
+        if DEBUG:
+            print("got job statuses", flush=True)
 
-        n_submitted = 0
         if DEBUG:
             print("%d: looping for %d subids" % (nanny_id, len(subids)), flush=True)
         for subid in subids:
@@ -143,50 +146,47 @@ def _nanny_function(
                         fut.cjob = cjob
                         exec._nanny_subids[nanny_id][subid] = (cjob, fut, None)
 
-                n_submitted += 1
-                if n_submitted >= 100:
-                    break
-                else:
-                    continue
+            if subid in statuses:
+                status_code = statuses[subid]
+                if status_code in ["4", "3", "5", "7"]:
+                    outfile = os.path.abspath(
+                        os.path.join(exec.execdir, subid, "output.pkl"))
+                    infile = os.path.abspath(
+                        os.path.join(exec.execdir, subid, "input.pkl"))
 
-            status_code = statuses[subid]
-            if status_code in ["4", "3", "5", "7"]:
-                outfile = os.path.abspath(
-                    os.path.join(exec.execdir, subid, "output.pkl"))
-                infile = os.path.abspath(
-                    os.path.join(exec.execdir, subid, "input.pkl"))
-
-                del ALL_CONDOR_JOBS[cjob]
-                subprocess.run(
-                    "condor_rm %s; condor_rm -forcex %s" % (cjob, cjob),
-                    shell=True,
-                    check=True,
-                    capture_output=True,
-                )
-                if os.path.exists(outfile):
-                    res = joblib.load(outfile)
-                elif status_code in ["3", "5", "7"]:
-                    res = RuntimeError(
-                        "Condor job %s: status %s" % (subid, STATUS_DICT[status_code])
+                    del ALL_CONDOR_JOBS[cjob]
+                    subprocess.run(
+                        "condor_rm %s; condor_rm -forcex %s" % (cjob, cjob),
+                        shell=True,
+                        check=True,
+                        capture_output=True,
                     )
-                else:
-                    res = RuntimeError(
-                        "Condor job %s: no status or job output found!" % subid)
+                    if os.path.exists(outfile):
+                        res = joblib.load(outfile)
+                    elif status_code in ["3", "5", "7"]:
+                        res = RuntimeError(
+                            "Condor job %s: status %s" % (
+                                subid, STATUS_DICT[status_code]
+                            )
+                        )
+                    else:
+                        res = RuntimeError(
+                            "Condor job %s: no status or job output found!" % subid)
 
-                subprocess.run(
-                    "rm -f %s %s" % (infile, outfile),
-                    shell=True,
-                    check=True,
-                )
+                    subprocess.run(
+                        "rm -f %s %s" % (infile, outfile),
+                        shell=True,
+                        check=True,
+                    )
 
-                if isinstance(res, Exception):
-                    fut.set_exception(res)
-                else:
-                    fut.set_result(res)
+                    if isinstance(res, Exception):
+                        fut.set_exception(res)
+                    else:
+                        fut.set_result(res)
 
-                del exec._nanny_subids[nanny_id][subid]
-                with ACTIVE_THREAD_LOCK:
-                    exec._num_jobs -= 1
+                    del exec._nanny_subids[nanny_id][subid]
+                    with ACTIVE_THREAD_LOCK:
+                        exec._num_jobs -= 1
 
 
 class CondorExecutor():
